high contrast mode implementation
