+++ src/interfaces.ts
@@ -1,0 +1,115 @@
+/*
+ *  Power BI Visualizations
+ *
+ *  Copyright (c) Microsoft Corporation
+ *  All rights reserved.
+ *  MIT License
+ *
+ *  Permission is hereby granted, free of charge, to any person obtaining a copy
+ *  of this software and associated documentation files (the ""Software""), to deal
+ *  in the Software without restriction, including without limitation the rights
+ *  to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ *  copies of the Software, and to permit persons to whom the Software is
+ *  furnished to do so, subject to the following conditions:
+ *
+ *  The above copyright notice and this permission notice shall be included in
+ *  all copies or substantial portions of the Software.
+ *
+ *  THE SOFTWARE IS PROVIDED *AS IS*, WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ *  IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ *  FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ *  AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ *  LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ *  OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
+ *  THE SOFTWARE.
+ */
+
+module powerbi.extensibility.visual {
+    // powerbi.data
+    import ISQExpr = powerbi.data.ISQExpr;
+
+    // powerbi.extensibility.utils.svg
+    import IMargin = powerbi.extensibility.utils.svg.IMargin;
+
+    // powerbi.extensibility.utils.interactivity
+    import SelectableDataPoint = powerbi.extensibility.utils.interactivity.SelectableDataPoint;
+
+    export interface ChicletSlicerData {
+        categorySourceName: string;
+        formatString: string;
+        slicerDataPoints: ChicletSlicerDataPoint[];
+        slicerSettings: ChicletSlicerSettings;
+        hasSelectionOverride?: boolean;
+        hasHighlights: boolean;
+        identityFields: ISQExpr[];
+    }
+
+    export interface ChicletSlicerDataPoint extends SelectableDataPoint {
+        category?: string;
+        value?: number;
+        mouseOver?: boolean;
+        mouseOut?: boolean;
+        isSelectAllDataPoint?: boolean;
+        imageURL?: string;
+        selectable?: boolean;
+        filtered?: boolean;
+    }
+
+    export interface ChicletSlicerSettings {
+        general: {
+            orientation: string;
+            columns: number;
+            rows: number;
+            multiselect: boolean;
+            forcedSelection: boolean;
+            showDisabled: string;
+            selection: string;
+            filter: any;
+            selfFilterEnabled: boolean;
+        };
+        margin: IMargin;
+        header: {
+            borderBottomWidth: number;
+            show: boolean;
+            outline: string;
+            fontColor: string;
+            background?: string;
+            textSize: number;
+            outlineColor: string;
+            outlineWeight: number;
+            title: string;
+        };
+        headerText: {
+            marginLeft: number;
+            marginTop: number;
+        };
+        slicerText: {
+            textSize: number;
+            height: number;
+            width: number;
+            fontColor: string;
+            selectedColor: string;
+            hoverColor: string;
+            unselectedColor: string;
+            disabledColor: string;
+            marginLeft: number;
+            outline: string;
+            background?: string;
+            transparency: number;
+            outlineColor: string;
+            outlineWeight: number;
+            padding: number;
+            borderStyle: string;
+        };
+        slicerItemContainer: {
+            marginTop: number;
+            marginLeft: number;
+        };
+        images: {
+            imageSplit: number;
+            imageRound: boolean;
+            stretchImage: boolean;
+            bottomImage: boolean;
+        };
+    }
+}
